+++ Examples/fortran_interface.h
@@ -1,3 +1,3 @@
-extern void write_int32(char*, long*);
+extern void write_int32(char*, int_least32_t*);
 
-extern void read_int32(char*, long*);
+extern void read_int32(char*, int_least32_t*);
