+++ Examples/fortran_interface.hpp
@@ -1,3 +1,3 @@
-extern "C" void write_int32(char*, long*);
+extern "C" void write_int32(char*, int_least32_t*);
 
-extern "C" void read_int32(char*, long*);
+extern "C" void read_int32(char*, int_least32_t*);
